+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to this project will be documented in this file.
 - *(items)* Added mobile screen for load items card
 - *(items)* Added mobile screen for attach pdf
 - *(cards)* Saved locally on browser encrypted
+- *(cards)* Fixed the load from cache and class refactor redundance
 
 ### ⚙️ Miscellaneous Tasks